move setStateString into builder
